[JT-60] style: package 명 변경
- paymentInfo -> payment
