+++ module-domain/src/main/java/com/devtoon/jtoon/payment/repository/MemberCookieRepository.java
@@ -1,8 +1,8 @@
-package com.devtoon.jtoon.paymentinfo.repository;
+package com.devtoon.jtoon.payment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.devtoon.jtoon.paymentinfo.entity.MemberCookie;
+import com.devtoon.jtoon.payment.entity.MemberCookie;
 
 public interface MemberCookieRepository extends JpaRepository<MemberCookie, Long> {
 }
