+++ module-domain/src/main/java/com/devtoon/jtoon/payment/repository/PaymentInfoRepository.java
@@ -1,9 +1,9 @@
-package com.devtoon.jtoon.paymentinfo.repository;
+package com.devtoon.jtoon.payment.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.devtoon.jtoon.paymentinfo.entity.PaymentInfo;
+import com.devtoon.jtoon.payment.entity.PaymentInfo;
 
 @Repository
 public interface PaymentInfoRepository extends JpaRepository<PaymentInfo, Long> {
